\name{KemansToGmm-package}
\alias{KemansToGmm-package}
\alias{KemansToGmm}
\docType{package}
\title{
\packageTitle{KemansToGmm}
}
\description{
\packageDescription{KemansToGmm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KemansToGmm}
\packageIndices{KemansToGmm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{KemansToGmm}

Maintainer: \packageMaintainer{KemansToGmm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
