% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KemansToGmmR.R
\name{kmtogmm}
\alias{kmtogmm}
\title{Firstly, the Kmeans algorithm is iterated to calculate the center of gravity of the class, and then the Gaussian mixture model is iterated to classify the samples.}
\usage{
kmtogmm(samp, g)
}
\arguments{
\item{samp}{the sample}

\item{g}{the Number of classes}

\item{pie}{the prior or posterior probabilities of all classifications}

\item{prob}{the probabilities of samples belong to all classifications}
}
\value{
the mean and variance of each class and the Sample classification
}
\description{
Firstly, the Kmeans algorithm is iterated to calculate the center of gravity of the class, and then the Gaussian mixture model is iterated to classify the samples.
}
\examples{
library(KemansToGmm)
samp = matrix(c(8.6338,8.8988,3.604,14.1556,19.2995,18.2959,31.7925,32.8074,32.0885,148.6553,141.5995,147.4205,7.0472,1.1754,7.5018,12.3305,17.0575,17.2132,33.5934,37.5266,36.9793,153.7012,151.2712,152.8625), nrow = 12, ncol = 2)
g = 4
kmtogmm(samp,g)
}
